# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Discord Moderation Bot Admin Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guilds:
    get:
      operationId: listGuilds
      tags: [guilds]
      summary: List guilds the user manages that have the bot
      responses:
        "200":
          description: List of guilds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guild"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /guilds/{guildId}:
    get:
      operationId: getGuild
      tags: [guilds]
      summary: Get a specific guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guild details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildDetail"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /guilds/{guildId}/channels:
    get:
      operationId: listGuildChannels
      tags: [guilds]
      summary: List channels in a guild, optionally filtered by Discord channel type
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: integer
          description: "Discord channel type filter: 0=text, 2=voice, 5=announcement. Omit to get text+announcement."
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
...
  /guilds/{guildId}/logs:
    get:
      operationId: listGuildLogs
      tags: [guilds]
      summary: Get recent audit log entries for a guild (last 50)
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
  /guilds/{guildId}/moderation:
    get:
      operationId: listGuildModeration
      tags: [guilds]
      summary: Get moderation action history for a guild (bans, kicks, timeouts, warns)
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: Filter actions targeting a specific user ID
      responses:
        "200":
          description: List of moderation log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
  /guilds/{guildId}/warnings/{userId}:
    get:
      operationId: listUserWarnings
      tags: [guilds]
      summary: Get all warnings for a specific user in a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of warning log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
  /guilds/{guildId}/leveling-config:
    get:
      operationId: getLevelingConfig
      tags: [leveling]
      summary: Get the leveling configuration for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Leveling configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LevelingConfig"
    put:
      operationId: updateLevelingConfig
      tags: [leveling]
      summary: Update the leveling configuration for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LevelingConfigInput"
      responses:
        "200":
          description: Updated leveling configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LevelingConfig"
  /guilds/{guildId}/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leveling]
      summary: Get the XP leaderboard for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [text, voice]
          description: "Which leaderboard: text (default) or voice"
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
  /guilds/{guildId}/dashboard-access:
    get:
      operationId: listDashboardAccess
      tags: [guilds]
      summary: List roles/users explicitly granted web dashboard access (read-only; managed via bot command)
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard access entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DashboardAccessEntry"
  /guilds/{guildId}/members/search:
    get:
      operationId: searchGuildMembers
      tags: [leveling]
      summary: Search current guild members by username/nickname
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Matching members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberSearchResult"
  /guilds/{guildId}/members/{userId}/xp:
    get:
      operationId: getMemberXp
      tags: [leveling]
      summary: Get a member's current XP/level
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member XP/level
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberXp"
    put:
      operationId: updateMemberXp
      tags: [leveling]
      summary: Admin override - directly set a member's XP/level
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberXpInput"
      responses:
        "200":
          description: Updated member XP/level
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberXp"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Guild:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: ["string", "null"]
        iconUrl:
          type: ["string", "null"]
        memberCount:
          type: integer
        botPresent:
          type: boolean
      required:
        - id
        - name
        - memberCount
        - botPresent
    GuildDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: ["string", "null"]
        iconUrl:
          type: ["string", "null"]
        memberCount:
          type: integer
        description:
          type: ["string", "null"]
      required:
        - id
        - name
        - memberCount
    Channel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: integer
      required:
        - id
        - name
        - type
...
    LogEntry:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        eventType:
          type: string
        userId:
          type: ["string", "null"]
        targetId:
          type: ["string", "null"]
        targetDisplayName:
          type: string
          nullable: true
        targetAvatarUrl:
          type: string
          nullable: true
        actorDisplayName:
          type: string
          nullable: true
        description:
          type: string
        metadata:
          type: object
        createdAt:
          type: string
      required:
        - id
        - guildId
        - eventType
        - description
        - createdAt
    LevelingConfig:
      type: object
      properties:
        guildId:
          type: string
        textXpMin:
          type: integer
        textXpMax:
          type: integer
        textXpCooldown:
          type: integer
        voiceXpPerMinute:
          type: integer
        levelupChannelId:
          type: ["string", "null"]
        levelupMessageEnabled:
          type: boolean
      required:
        - guildId
        - textXpMin
        - textXpMax
        - textXpCooldown
        - voiceXpPerMinute
        - levelupMessageEnabled
    LevelingConfigInput:
      type: object
      properties:
        textXpMin:
          type: integer
        textXpMax:
          type: integer
        textXpCooldown:
          type: integer
        voiceXpPerMinute:
          type: integer
        levelupChannelId:
          type: ["string", "null"]
        levelupMessageEnabled:
          type: boolean
    LeaderboardEntry:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        xp:
          type: integer
        level:
          type: integer
        rank:
          type: integer
      required:
        - userId
        - xp
        - level
        - rank
...
    DashboardAccessEntry:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        targetType:
          type: string
          enum: [role, user]
        targetId:
          type: string
        addedBy:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - guildId
        - targetType
        - targetId
        - createdAt
    MemberSearchResult:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required:
        - userId
...